Created and implemented both shield and armor systems. Fixed a bug with 2 effects with different targetting both using manual targeting. Moving on to expanding saving and loading system for future card augmenting system

+++ Assets/Scripts/card system/model/StatMods.cs
@@ -5,10 +5,10 @@ namespace model.entity_state
     public class StatMods
     {
         public StatModType type;
-        public int amount;
+        public float amount;
         public int remainingTurns;
 
-        public StatMods(StatModType type, int amount, int remainingTurns)
+        public StatMods(StatModType type, float amount, int remainingTurns)
         {
             this.type = type;
             this.amount = amount;
